// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/commercial-location.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Commercial location information
 */
export interface CommercialLocation { 
    /**
     * Numerical identifier of the country.
     */
    countryCodeUic: string;
    /**
     * ISO code of the country.
     */
    countryCodeIso?: string;
    /**
     * Numerical identifier.
     */
    number: string;
    /**
     * Identifier of the owner.
     */
    owner: string;
    /**
     * name.
     */
    name?: string;
}