// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/freight-wagon-location.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FreightWagonLocation { 
    /**
     * Freight Location number.
     */
    number: string;
    /**
     * Satellite number of the station
     */
    satelliteNumber?: string;
    name?: string;
}