// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/order-summary.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Status } from './status';
import { NHM } from './nhm';
import { CustomerInformation } from './customer-information';
import { DemandWagonType } from './demand-wagon-type';
import { TransitRailwayUndertaking } from './transit-railway-undertaking';
import { DemandLocation } from './demand-location';
import { CommercialLocation } from './commercial-location';


/**
 * order summary
 */
export interface OrderSummary { 
    /**
     * ID of the order.
     */
    orderId?: string;
    /**
     * internal Id of the order in link2rail
     */
    orderIdConsumer?: string;
    /**
     * Internal Number of the order.
     */
    internalOrderNumber?: string;
    /**
     * Name of the template that was used for original order request.
     */
    templateName?: string;
    orderer?: CustomerInformation;
    shipper?: CustomerInformation;
    demandLocation?: DemandLocation;
    demandWagonType?: DemandWagonType;
    /**
     * Number of ordered wagons.
     */
    numberOfWagonsOrdered?: number;
    /**
     * Forecasted number of wagons which will be disposed.
     */
    numberOfWagonsDisposed?: any | null;
    status?: Status;
    /**
     * Origin of original order.
     */
    origin?: string;
    /**
     * Desired date and time of empty wagon delivery.
     */
    deliveryDateTime?: string;
    loadRunLocation?: CommercialLocation;
    transitRailwayUndertaking?: TransitRailwayUndertaking;
    nhm?: NHM;
    /**
     * Comment addressed to customer service.
     */
    commentCustomer?: string;
    /**
     * Comment addressed to customer service.
     */
    commentDisponent?: string;
    /**
     * Reference number for the orders chosen by the customer.
     */
    customerReference?: string;
    /**
     * Flag indicating whether the order can be updated
     */
    isEditable?: boolean;
    /**
     * Flag indicating whether the order can be canceled
     */
    isCancelable?: boolean;
}
export namespace OrderSummary {
}