// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/status.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type Status = 'CANCELED' | 'CREATED' | 'DISPATCHED' | 'DRAFT' | 'DRAFT_REMOVED' | 'IN_PROCESS' | 'IS_TRANSMITTED' | 'PARTIAL_CANCELLATION' | 'REJECTED' | 'TRANSMITTED' | 'TRANSMIT_FAILED';

export const Status = {
    Canceled: 'CANCELED' as Status,
    Created: 'CREATED' as Status,
    Dispatched: 'DISPATCHED' as Status,
    Draft: 'DRAFT' as Status,
    DraftRemoved: 'DRAFT_REMOVED' as Status,
    InProcess: 'IN_PROCESS' as Status,
    IsTransmitted: 'IS_TRANSMITTED' as Status,
    PartialCancellation: 'PARTIAL_CANCELLATION' as Status,
    Rejected: 'REJECTED' as Status,
    Transmitted: 'TRANSMITTED' as Status,
    TransmitFailed: 'TRANSMIT_FAILED' as Status
};